Удалил интерфейс AdsServiceImpl.java, поднял выше сервис AdsService
